lekinetworks.server/requirements.txt
remnawave
fastapi
uvicorn
python-dotenv
aiomysql
pyjwt
aiosmtplib
yookassa
httpx
